git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@6988 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/containers/Queue.h
@@ -28,7 +28,7 @@
 ===============================================================================
 */
 
-#define idQueue( type, next )		idQueueTemplate<type, (int)&(((type*)NULL)->next)>
+#define idQueue( type, next )		idQueueTemplate<type, (int)(size_t)&(((type*)NULL)->next)>
 
 template< class type, int nextOffset >
 class idQueueTemplate {
